+++ Hoa/Prototype/CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.16.01.11
+
+  * Quality: Run devtools:cs. (Ivan Enderlin, 2016-01-09T09:07:19+01:00)
+  * Consistency: Use `Hoa\Consistency`. (Ivan Enderlin, 2015-12-08T21:41:46+01:00)
+
 # 0.15.09.08
 
   * Add a `.gitignore` file. (Stéphane HULARD, 2015-08-03T11:42:58+02:00)